Login_vpn add dialog for delete user

+++ app/src/main/java/com/telephone/coursetable/Login_vpn.java
@@ -590,24 +590,39 @@ public class Login_vpn extends AppCompatActivity {
         final String NAME = "deleteUser()";
         clearIMAndFocus();
         String sid = ((AutoCompleteTextView) findViewById(R.id.sid_input)).getText().toString();
-         new Thread((Runnable) () -> {
+        getAlertDialog("确定要取消记住用户" + " " + sid + " " + "的登录信息吗？",
+                (DialogInterface.OnClickListener) (dialogInterface, i) -> new Thread((Runnable) () -> {
                     udao.deleteUser(sid);
                     Log.e(NAME + " " + "user deleted", sid);
                     updateUserNameAutoFill();
-
                     runOnUiThread((Runnable) () -> {
-                        ((AutoCompleteTextView) findViewById(R.id.sid_input)).setText("");
-                        ((AutoCompleteTextView) findViewById(R.id.passwd_input)).setText("");
-
                         aaw_pwd = "";
                         sys_pwd = "";
-
-                        setFocusToEditText((EditText) findViewById(R.id.sid_input));
+                        ((AutoCompleteTextView)findViewById(R.id.sid_input)).setText("");
+                        ((AutoCompleteTextView)findViewById(R.id.passwd_input)).setText("");
+                        setFocusToEditText((EditText)findViewById(R.id.sid_input));
                     });
-         }).start();
+                }).start(),
+                (DialogInterface.OnClickListener) (dialogInterface, i) -> {},
+                null, null).show();
 
     }
 
+    private AlertDialog getAlertDialog(@Nullable final String m, @NonNull DialogInterface.OnClickListener yes, @NonNull DialogInterface.OnClickListener no, @Nullable View view, @Nullable String title){
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        if (m != null) {
+            builder.setMessage(m);
+        }
+        builder.setPositiveButton(getResources().getString(R.string.ok_btn_text_zhcn), yes)
+                .setNegativeButton(getResources().getString(R.string.deny_btn_zhcn), no);
+        if (view != null){
+            builder.setView(view);
+        }
+        if (title != null){
+            builder.setTitle(title);
+        }
+        return builder.create();
+    }
 
     /**
      * @return - true : everything is ok
